requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
transformers
torch
huggingface-hub
accelerate
sentence-transformers
faiss-cpu